requirements.txt
python-dotenv>=1.0.0
SQLAlchemy>=2.0.0
PyYAML>=6.0
